+++ python/lsst/ts/mtdome/llc_configuration_limits/abstract_limits.py
@@ -33,7 +33,5 @@ class AbstractLimits(ABC):
     """
 
     @abstractmethod
-    def validate(
-        self, configuration_parameters: dict
-    ) -> typing.List[typing.Dict[str, typing.Any]]:
+    def validate(self, configuration_parameters: dict) -> list[dict[str, typing.Any]]:
         pass